updating package name
