+++ app/integration/runtime-springboot/src/main/java/io/syndesis/integration/runtime/sb/logging/IntegrationLoggingConfiguration.java
@@ -13,21 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.integration.runtime;
+package io.syndesis.integration.runtime.sb.logging;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-@ConfigurationProperties(prefix = "syndesis.integration.runtime")
-public class IntegrationRuntimeConfiguration {
+@ConfigurationProperties(prefix = "syndesis.integration.runtime.logging")
+public class IntegrationLoggingConfiguration {
     /**
-     * Enable/Disable syndesis runtime.
+     * Enable/Disable syndesis runtime logging.
      */
     private boolean enabled = true;
-
-    /**
-     * The Syndesis configuration location;
-     */
-    private String configurationLocation = "classpath:syndesis/integration/integration.json";
 
     public boolean isEnabled() {
         return enabled;
@@ -35,13 +30,5 @@ public class IntegrationRuntimeConfiguration {
 
     public void setEnabled(boolean enabled) {
         this.enabled = enabled;
-    }
-
-    public String getConfigurationLocation() {
-        return configurationLocation;
-    }
-
-    public void setConfigurationLocation(String configurationLocation) {
-        this.configurationLocation = configurationLocation;
     }
 }
