+++ app/integration/runtime-springboot/src/test/java/io/syndesis/integration/component/proxy/ComponentProxyCustomizerAddPropertyTest.java
@@ -17,7 +17,6 @@ package io.syndesis.integration.component.proxy;
 
 import java.util.Map;
 
-import io.syndesis.integration.runtime.IntegrationRuntimeAutoConfiguration;
 import org.apache.camel.Body;
 import org.apache.camel.CamelContext;
 import org.apache.camel.ProducerTemplate;
@@ -32,6 +31,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringRunner;
+
+import io.syndesis.integration.runtime.sb.IntegrationRuntimeAutoConfiguration;
 
 import static org.assertj.core.api.Assertions.assertThat;
